+++ src/main/java/searchengine/dto/search/SearchResponse.java
@@ -4,8 +4,6 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 
 import java.util.List;
 
@@ -22,9 +20,5 @@ public class SearchResponse {
 	public SearchResponse(boolean result, String error) {
 		this.result = result;
 		this.error = error;
-	}
-
-	public ResponseEntity<SearchResponse> emptyQuery() {
-		return new ResponseEntity<>(new SearchResponse(false, "Empty query"), HttpStatus.BAD_REQUEST);
 	}
 }