+++ src/main/java/searchengine/exceptionhandler/EmptyQueryException.java
@@ -1,7 +1,0 @@
-package searchengine.exceptionhandler;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-
-public class EmptyQueryException extends RuntimeException{
-}