+++ src/main/java/searchengine/model/PageEntity.java
@@ -18,36 +18,21 @@ import java.util.Set;
 public class PageEntity {
 
 	@Id
+	@Column(name = "id", nullable = false)
 	@SequenceGenerator(
 			name = "page_seq",
 			sequenceName = "page_sequence",
-			initialValue = 1, allocationSize = 50)
+			allocationSize = 40)
 	@GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "page_seq")
-	@Column(name = "id", nullable = false)
 	private Integer id;
 
-	//	cascade = {CascadeType.REMOVE, CascadeType.MERGE, CascadeType.REFRESH}
 	@ManyToOne(fetch = FetchType.LAZY, targetEntity = SiteEntity.class, cascade = {CascadeType.MERGE, CascadeType.REFRESH}, optional = false)
 	@OnDelete(action = OnDeleteAction.NO_ACTION)
 	@JoinColumn(foreignKey = @ForeignKey(name = "site_page_FK"), columnDefinition = "Integer",
 			referencedColumnName = "id", name = "site_id", nullable = false, updatable = false)
 	private SiteEntity siteEntity;
 
-	/*
-	https://www.navicat.com/en/company/aboutus/blog/1308-choosing-between-varchar-and-text-in-mysql#:~:text=Some%20Differences%20Between%20VARCHAR%20and%20TEXT&text=The%20VAR%20in%20VARCHAR%20means,be%20part%20of%20an%20index.
-	 The max length of a varchar is subject to the max row size in MySQL, which is 64KB (not counting BLOBs):
-	VARCHAR(65535)
-
-	However, note that the limit is lower if you use a multi-byte character set:
-	VARCHAR(21844) CHARACTER SET utf8
-
-	15.22 InnoDB Limits
-	https://dev.mysql.com/doc/refman/8.0/en/innodb-limits.html
-
-	 */
-
 	@Column(name = "path", columnDefinition = "VARCHAR(768) CHARACTER SET utf8")
-//	@Column(name = "path", columnDefinition = "TEXT NOT NULL, KEY path_index (path(255))")
 	private String path;
 
 	@Column(nullable = false)
@@ -57,13 +42,10 @@ public class PageEntity {
 	private String content;
 
 	@ManyToMany(fetch = FetchType.LAZY)
-//	@OnDelete(action = OnDeleteAction.CASCADE)
 	@JoinTable(
 			name = "search_index",
 			joinColumns = {@JoinColumn(name = "page_id")},
 			inverseJoinColumns = {@JoinColumn(name = "lemma_id")})
-//	@ManyToMany(fetch = FetchType.LAZY, mappedBy = "pageEntities")
-//	@OnDelete(action = OnDeleteAction.CASCADE)
 	private Set<LemmaEntity> lemmaEntities = new HashSet<>();
 
 	public PageEntity(SiteEntity siteEntity, int code, String content, String path) {