+++ src/main/java/searchengine/repositories/SiteRepository.java
@@ -9,14 +9,9 @@ import searchengine.model.SiteEntity;
 @Repository
 public interface SiteRepository extends JpaRepository<SiteEntity, Long> {
 
-	void deleteById(Integer id);
-
 	boolean existsByUrl(String url);
 
 	SiteEntity findByUrl(String url);
-
-	SiteEntity findById(Integer id);
-
 
 }