+++ src/main/java/searchengine/services/Impl/SearchServiceImpl.java
@@ -7,7 +7,6 @@ import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 import org.jsoup.Jsoup;
 import org.jsoup.nodes.Document;
-import org.springframework.data.domain.*;
 import org.springframework.stereotype.Service;
 import searchengine.dto.search.SearchData;
 import searchengine.dto.search.SearchResponse;
@@ -15,10 +14,7 @@ import searchengine.model.IndexEntity;
 import searchengine.model.LemmaEntity;
 import searchengine.model.PageEntity;
 import searchengine.model.SiteEntity;
-import searchengine.repositories.IndexRepository;
-import searchengine.repositories.LemmaRepository;
-import searchengine.repositories.PageRepository;
-import searchengine.repositories.SiteRepository;
+import searchengine.services.RepositoryService;
 import searchengine.services.SearchService;
 import searchengine.tools.LemmaFinder;
 import searchengine.tools.SnippetGenerator;
@@ -32,27 +28,20 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class SearchServiceImpl implements SearchService {
 
-	public static final double ratio = 1.80;
-	private final SiteRepository siteRepository;
-	private final PageRepository pageRepository;
-	private final LemmaRepository lemmaRepository;
-	private final IndexRepository indexRepository;
+	public static final double ratio = 1.85;
+	private String rarestLemma = null;
 	private final LemmaFinder lemmaFinder;
 	private final SnippetGenerator snippetGenerator;
+	private final RepositoryService repositoryService;
 
-	private String rarestLemma = null;
-	private Map<String, LemmaEntity> retrievedLemmas = new HashMap<>();
+	private List<PageEntity> pagesByRarestLemma = new ArrayList<>();
 	private Map<String, LemmaEntity> nextStepLemmas = new HashMap<>();
-	private List<PageEntity> currentPageEntities = new ArrayList<>();
 	private Map<PageEntity, Float> totalPagesWithRelevance = new HashMap<>();
-	private List<PageEntity> totalPages = new ArrayList<>();
 
 	@Override
 	@Builder
 	public SearchResponse getSearchResults(@NotNull String query, String siteUrl, Integer offset, Integer limit) {
 		if (query.isEmpty()) return emptyQuery();
-
-		int totalPagesCount = 0;
 
 		//Get lemmas from search query
 		Set<String> queryLemmas = lemmaFinder.getLemmaSet(query);
@@ -60,21 +49,21 @@ public class SearchServiceImpl implements SearchService {
 
 		//One or all sites
 		List<SiteEntity> siteEntities = siteUrl != null
-				? Collections.singletonList(siteRepository.findByUrl(siteUrl))
-				: siteRepository.findAll();
+				? Collections.singletonList(repositoryService.getSiteByUrl(siteUrl))
+				: repositoryService.getSites();
+
+		int totalPagesCount = 0;
 
 		for (SiteEntity site : siteEntities) {
-			//Get lemmas from DB by Site
-			retrievedLemmas = getLemmaEntitiesFromTable(queryLemmas, site);
+			Map<String, LemmaEntity> retrievedLemmas = getLemmaEntitiesFromTable(queryLemmas, site);
 			if (retrievedLemmas.size() != 0) {
 
 				//Calculate threshold of frequency for deleting
 				int totalFreq = getTotalFrequency(retrievedLemmas);
 				float thresholdFreq = (float) ((totalFreq / retrievedLemmas.size()) * ratio);
 
-				//Delete most frequently lemmas
+				//Delete most frequently lemmas, Sort by frequency
 				nextStepLemmas = removeMostFrequentlyLemmas(retrievedLemmas, thresholdFreq);
-				//Sort by frequency
 				nextStepLemmas = sortByFrequency(nextStepLemmas);
 
 				//Get list of Pages by rarest lemma
@@ -82,37 +71,19 @@ public class SearchServiceImpl implements SearchService {
 				if (rarestLemma != null) {
 					List<Integer> pageIds = getPageEntitiesIdsByRarestLemma(site);
 
-					if (siteEntities.size() == 1) {
-						List<PageEntity> pagesList = pageRepository.findAllByIdIn(pageIds);
-//						List<PageEntity> subPages = pagesList.stream()
-//								.skip(offset)
-//								.limit(limit)
-//								.toList();
-						Pageable pageable = PageRequest.of(offset, limit);
-						Page<PageEntity> pageableObject = pageRepository.findAllByIdIn(pageIds, pageable);
-//						List<PageEntity> onlyLimitPages = pageableObject.getContent();
-
-//						Page<PageEntity> currentPageObj = new PageImpl<>(subPages, PageRequest.of(offset, limit), totalPages);
-						currentPageEntities = pageableObject.getContent();
-						totalPagesCount = totalPagesCount + currentPageEntities.size();
-					} else {
-						currentPageEntities = pageRepository.findAllByIdIn(pageIds);
-						totalPagesCount = totalPagesCount + currentPageEntities.size();
-					}
+					pagesByRarestLemma = repositoryService.getPagesByIds(pageIds);
+					totalPagesCount += pagesByRarestLemma.size();
 				}
-
-//				printRarestLemma(site, rarestLemma, new HashSet<>(currentPageEntities));
 			}
 
 			//We continue to look for the remaining lemmas on these pages
 			Map<String, LemmaEntity> finalLemmas = getFinalSortedLemmasMap(rarestLemma, nextStepLemmas);
 
-			List<PageEntity> finalPages = new ArrayList<>(currentPageEntities);
+			List<PageEntity> finalPages = new ArrayList<>(pagesByRarestLemma);
 			for (String desiredLemma : finalLemmas.keySet()) {
-				List<PageEntity> eachIterationPages = getRetainedPages(site, currentPageEntities, desiredLemma);
+				List<PageEntity> eachIterationPages = getRetainedPages(site, pagesByRarestLemma, desiredLemma);
 				finalPages.clear();
 				finalPages.addAll(eachIterationPages);
-				totalPagesCount = totalPagesCount + finalPages.size();
 			}
 
 			//Calculate absolute relevancy
@@ -120,35 +91,17 @@ public class SearchServiceImpl implements SearchService {
 			totalPagesWithRelevance.putAll(pagesWithAbsRelativeMap);
 		}
 
-		System.out.println("before calculate relevance");
-		for (PageEntity pp:totalPagesWithRelevance.keySet()) {
-			System.out.println(pp.getId() + " " + getTitle(pp.getContent()) + " absolute relevance = " + totalPagesWithRelevance.get(pp));
-		}
-		totalPagesWithRelevance = calculateRelevance(totalPagesWithRelevance, getMaxRank(totalPagesWithRelevance));
-		System.out.println("after calculate relevance");
-		for (PageEntity pp:totalPagesWithRelevance.keySet()) {
-			System.out.println(pp.getId() + " " + getTitle(pp.getContent()) + " absolute relevance = " + totalPagesWithRelevance.get(pp));
-		}
-
+		float maxRank = getMaxRank(totalPagesWithRelevance);
+		totalPagesWithRelevance = calculateRelevance(totalPagesWithRelevance, maxRank);
 		totalPagesWithRelevance = sortByRank(totalPagesWithRelevance);
-		System.out.println("after sort");
-		for (PageEntity pp:totalPagesWithRelevance.keySet()) {
-			System.out.println(pp.getId() + " " + getTitle(pp.getContent()) + " absolute relevance = " + totalPagesWithRelevance.get(pp));
-		}
+
+		//Consider limit offset
+		Map<PageEntity, Float> subsetMap = getSortedSubsetMap(totalPagesWithRelevance, limit, offset);
 
 		List<SearchData> totalData = new ArrayList<>();
 
-		Map<PageEntity, Float> subsetMap = new HashMap<>();
-		if (siteEntities.size() == 1) {
-			subsetMap = new HashMap<>(totalPagesWithRelevance);
-			subsetMap = sortByRank(subsetMap);
-		} else if (siteEntities.size() > 1) {
-			subsetMap = getSortedSubsetMap(totalPagesWithRelevance, limit, offset);
-			totalPagesCount = totalPagesWithRelevance.size();
-		}
-		totalPagesWithRelevance.clear();
-
 		for (PageEntity page : subsetMap.keySet()) {
+
 			SearchData searchData = SearchData.builder()
 					.uri(page.getPath())
 					.siteName(page.getSiteEntity().getName())
@@ -156,14 +109,12 @@ public class SearchServiceImpl implements SearchService {
 					.snippet(getSnippet(page, queryList))
 					.relevance(subsetMap.get(page))
 					.title(getTitle(page.getContent())).build();
+
 			totalData.add(searchData);
 		}
-		if (totalPagesCount == 0) {
-			totalData = new ArrayList<>();
-		}
-		if (totalData.size() == 0) {
-			totalPagesCount = 0;
-		}
+
+		totalPagesWithRelevance.clear();
+
 		return SearchResponse.builder()
 				.result(true)
 				.count(totalPagesCount)
@@ -181,15 +132,15 @@ public class SearchServiceImpl implements SearchService {
 	}
 
 	public Map<PageEntity, Float> getSortedSubsetMap(@NotNull Map<PageEntity, Float> source, int limit, int offset) {
-		// Сортируем Map в порядке убывания значений Float
+		//Sort Map in descending order of Float values
 		List<Map.Entry<PageEntity, Float>> sortedEntries = new ArrayList<>(source.entrySet());
 		sortedEntries.sort(Map.Entry.comparingByValue(Comparator.reverseOrder()));
 
-		// Выбираем подмножество элементов
+		//Select a subset of elements
 		int endIndex = Math.min(offset + limit, source.size());
 		List<Map.Entry<PageEntity, Float>> subsetEntries = sortedEntries.subList(offset, endIndex);
 
-		// Создаем новую Map с выбранными элементами
+		//Create a new Map with the selected elements
 		Map<PageEntity, Float> result = new LinkedHashMap<>();
 		for (Map.Entry<PageEntity, Float> entry : subsetEntries) {
 			result.put(entry.getKey(), entry.getValue());
@@ -199,7 +150,7 @@ public class SearchServiceImpl implements SearchService {
 
 	private @NotNull Map<String, LemmaEntity> getLemmaEntitiesFromTable(@NotNull Set<String> queryLemmas, SiteEntity site) {
 		return queryLemmas.stream()
-				.map(lemma -> lemmaRepository.findByLemmaAndSiteEntity(lemma, site))
+				.map(lemma -> repositoryService.getLemmaByNameFromSite(lemma, site))
 				.filter(Objects::nonNull)
 				.collect(Collectors.toMap(LemmaEntity::getLemma, Function.identity()));
 	}
@@ -219,44 +170,32 @@ public class SearchServiceImpl implements SearchService {
 
 	private Map<String, LemmaEntity> sortByFrequency(@NotNull Map<String, LemmaEntity> lemmaEntities) {
 		List<Map.Entry<String, LemmaEntity>> list = new ArrayList<>(lemmaEntities.entrySet());
-		list.sort((o1, o2) -> o1.getValue().getFrequency() - o2.getValue().getFrequency());
+		list.sort(Comparator.comparingInt(o -> o.getValue().getFrequency()));
 		return list.stream()
 				.collect(Collectors.toMap(Map.Entry::getKey, Map.Entry::getValue,
 						(oldValue, newValue) -> oldValue, LinkedHashMap::new));
 	}
 
 	private @NotNull List<PageEntity> getRetainedPages(SiteEntity site, @NotNull List<PageEntity> finalPages, String lemma) {
-		LemmaEntity lemmaEntity = lemmaRepository.findByLemmaAndSiteEntity(lemma, site);
-		Set<IndexEntity> indexSet = indexRepository.findAllByLemmaEntity(lemmaEntity);
+		LemmaEntity lemmaEntity = repositoryService.getLemmaByNameFromSite(lemma, site);
+		Set<IndexEntity> indexSet = repositoryService.getAllByLemma(lemmaEntity);
 		return finalPages.stream()
 				.filter(pageEntity -> indexSet.stream()
 						.anyMatch(index -> Objects.equals(index.getPageEntity().getId(), pageEntity.getId())))
 				.toList();
 	}
 
-	private Map<PageEntity, Float> getPagesWithRelevance(@NotNull List<PageEntity> source) {
-		if (source.size() != 0) {
-			Map<PageEntity, Float> pagesAbsoluteMap = calculateAbsRelevance(source, retrievedLemmas);
-			float maxAbsRel = getMaxRank(pagesAbsoluteMap);
-			return calculateRelevance(pagesAbsoluteMap, maxAbsRel);
-		}
-		return new HashMap<>();
-	}
-
 	private @NotNull Map<PageEntity, Float> calculateAbsRelevance(@NotNull List<PageEntity> source, Map<String, LemmaEntity> lemmas) {
 		Map<PageEntity, Float> result = new HashMap<>();
 		for (PageEntity p : source) {
-			System.out.println("page " + p.getId() + " " + getTitle(p.getContent()));
 			float absRelPage = (float) 0;
 			for (LemmaEntity l : lemmas.values()) {
-				IndexEntity idx = indexRepository.findByLemmaEntityAndPageEntity(l, p);
+				IndexEntity idx = repositoryService.getIndexByLemmaFromPage(l, p);
 				if (idx != null) {
-					System.out.println("lemma " + l.getLemma() + " " + l.getId() + " exists");
 					absRelPage += idx.getLemmaRank();
 				}
 			}
 			result.put(p, absRelPage);
-			System.out.println("page " + p.getId() + " " + getTitle(p.getContent()) + " has absolute rel = " + absRelPage);
 		}
 		return result;
 	}
@@ -267,7 +206,6 @@ public class SearchServiceImpl implements SearchService {
 	}
 
 	private Map<PageEntity, Float> calculateRelevance(@NotNull Map<PageEntity, Float> source, float maxAbsRel) {
-		System.out.println("Max relevance = " + maxAbsRel);
 		return source.keySet()
 				.stream()
 				.collect(Collectors.toMap(p -> p, p -> source.get(p) / maxAbsRel, (a, b) -> b));
@@ -291,31 +229,16 @@ public class SearchServiceImpl implements SearchService {
 		return snippetGenerator.generateSnippets();
 	}
 
-
-//	private void printResultPages(String results_pages, Set<PageEntity> resultPages) {
-//		System.out.println(results_pages);
-//		resultPages.forEach(x -> {
-//			System.out.println(x + " " + x.getPath());
-//		});
-//		System.out.println("-----------------------------------------------------------");
-//	}
-
-//	private void printRarestLemma(SiteEntity site, String rarestLemma, Set<PageEntity> rarestPages) {
-//		if (rarestLemma != null && rarestPages != null) {
-//			System.out.println("Rarest lemma - " + rarestLemma + " freq - " + lemmaRepository.getByLemmaAndSiteEntity(rarestLemma, site).getFrequency());
-//			printResultPages("pages by rarest lemma", rarestPages);
-//		}
-//	}
-
 	private @NotNull String getTitle(String sourceHtml) {
 		Document doc = Jsoup.parse(sourceHtml);
 		return doc.title();
 	}
 
 	private @Nullable Set<IndexEntity> getRarestIndexSet(SiteEntity site, String rarestLemma) {
-		LemmaEntity rarestLemmaEntity = lemmaRepository.findByLemmaAndSiteEntity(rarestLemma, site);
+		LemmaEntity rarestLemmaEntity = repositoryService.getLemmaByNameFromSite(rarestLemma, site);
 		if (rarestLemmaEntity != null)
-			return indexRepository.findAllByLemmaEntity(rarestLemmaEntity);
+			return repositoryService.getAllByLemma(rarestLemmaEntity);
+
 		return null;
 	}