+++ src/main/java/searchengine/services/PagesSavingService.java
@@ -2,7 +2,6 @@ package searchengine.services;
 
 import searchengine.model.PageEntity;
 import searchengine.model.SiteEntity;
-import searchengine.tools.StringPool;
 
 import java.util.concurrent.BlockingQueue;
 
@@ -12,7 +11,6 @@ public interface PagesSavingService {
 	void setIncomeQueue(BlockingQueue<PageEntity> incomeQueue);
 	void setOutcomeQueue(BlockingQueue<Integer> outcomeQueue);
 	void setSiteEntity(SiteEntity siteEntity);
-	void setStringPool(StringPool stringPool);
 }