+++ src/main/java/searchengine/services/StatisticsService.java
@@ -1,6 +1,5 @@
 package searchengine.services;
 
-import lombok.extern.slf4j.Slf4j;
 import searchengine.dto.statistics.StatisticsResponse;
 
 public interface StatisticsService {