+++ src/main/java/searchengine/tools/Regex.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public class Regex {
 	public static final String URL_IS_FILE_LINK = "https?:/(?:/[^/]+)+/[А-Яа-яёЁ\\w\\W ]+\\.[a-z]{3,5}(?!/|[\\wА-Яа-яёЁ])";
-	public static final String URL_IS_VALID = "^(ht|f)tp(s?)://[0-9a-zA-Z]([-.\\w]*[0-9a-zA-Z])*(:(0-9)*)*(/?)([a-zA-Z0-9\\-.?,'/\\\\+&%_]*)?$";
+	public static final String URL_IS_VALID = "^(ht|f)tp(s?)://[0-9a-zA-Z]([-.\\w]*[0-9a-zA-Z])*(:(0-9)*)*(/?)([a-zA-Z0-9\\-.?,'=/\\\\+%_]*)?$";
 
 	public static final List<String> HTML_EXT = new ArrayList<>() {{
 		add("html");