+++ src/main/java/searchengine/tools/SnippetGenerator.java
@@ -8,7 +8,6 @@ import org.jetbrains.annotations.NotNull;
 import org.jsoup.Jsoup;
 import org.jsoup.safety.Safelist;
 import org.springframework.stereotype.Component;
-import searchengine.tools.LemmaFinder;
 
 import java.util.*;
 import java.util.stream.Collectors;